% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_draws.R
\name{calc_stats_for_draws_predictions}
\alias{calc_stats_for_draws_predictions}
\title{Calculates statistics on draws from posterior for predictions (e.g.
concentrations)}
\usage{
calc_stats_for_draws_predictions(post, obs_type = "pk")
}
\arguments{
\item{post}{posterior object}

\item{obs_type}{observation type label, default is \code{pk}}
}
\value{
data frame with statistics for each observation timepoint, such as
mean, median, sd, confidence intervals etc.
}
\description{
Calculates statistics on draws from posterior for predictions (e.g.
concentrations)
}
\keyword{internal}
