% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_input_data}
\alias{check_input_data}
\title{Check that the supplied input data is OK for the model}
\usage{
check_input_data(mod, data)
}
\arguments{
\item{mod}{compiled Stan model}

\item{data}{dataset (see \code{\link[=new_stan_data]{new_stan_data()}})}
}
\description{
Check that the supplied input data is OK for the model
}
