% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stan_input_data.R
\name{check_stan_input_data}
\alias{check_stan_input_data}
\title{Check that the supplied input data for Stan is OK for the model}
\usage{
check_stan_input_data(code, data)
}
\arguments{
\item{code}{Stan model code}

\item{data}{input data, generated using \code{new_stan_data()}, which includes
an element \code{stan_data} that is checked in this function.}
}
\description{
Check that the supplied input data for Stan is OK for the model
}
