% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_draws.R
\name{extract_from_draws}
\alias{extract_from_draws}
\title{Extract info for observed data from posterior draws.
Includes e.g. the quantile of the observed data in the posterior draws
and the mean, 95\% CI, and standard deviations of the posterior.}
\usage{
extract_from_draws(post, data, filter = "ipred_obs", verbose = TRUE)
}
\arguments{
\item{post}{posterior object from CmdStanR}

\item{data}{data.frame with observed data}

\item{filter}{column name filter to use}

\item{verbose}{verbosity}
}
\description{
Extract info for observed data from posterior draws.
Includes e.g. the quantile of the observed data in the posterior draws
and the mean, 95\% CI, and standard deviations of the posterior.
}
