% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior_observation_info.R
\name{extract_posterior_observation_info}
\alias{extract_posterior_observation_info}
\title{Extract info for observed data from posterior draws.
Includes e.g. the quantile of the observed data in the posterior draws
and the meanm, 95\% CI, and standard deviations of the posterior.}
\usage{
extract_posterior_observation_info(post, data, verbose = TRUE)
}
\arguments{
\item{post}{posterior object from CmdStanR}

\item{data}{data.frame with observed data}

\item{verbose}{verbosity}
}
\description{
Extract info for observed data from posterior draws.
Includes e.g. the quantile of the observed data in the posterior draws
and the meanm, 95\% CI, and standard deviations of the posterior.
}
