% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stan_code.R
\name{generate_stan_code}
\alias{generate_stan_code}
\title{Generate a Stan model code from model definitions and template code.
This is an internal function. For creating a Stan model, please use
\code{new_stan_model()}.}
\usage{
generate_stan_code(template_code, definitions)
}
\arguments{
\item{template_code}{template model code}

\item{definitions}{list of model definitions, parsed using
\code{parse_model_definitions()}.}
}
\description{
Generate a Stan model code from model definitions and template code.
This is an internal function. For creating a Stan model, please use
\code{new_stan_model()}.
}
