% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stan_model.R
\name{generate_stan_model}
\alias{generate_stan_model}
\title{Generate a Stan model from model definitions and template code.
This is an internal function. For creating a Stan model, please use
\code{new_stan_model()}.}
\usage{
generate_stan_model(template_code, definitions)
}
\arguments{
\item{template_code}{template model code}

\item{definitions}{list of model definitions, parsed using
\code{parse_model_definitions()}.}
}
\description{
Generate a Stan model from model definitions and template code.
This is an internal function. For creating a Stan model, please use
\code{new_stan_model()}.
}
