% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_init.R
\name{get_init}
\alias{get_init}
\title{Get prior for Torsten from a PKPDsim model package}
\usage{
get_init(model, map = list(), drop = c())
}
\arguments{
\item{model}{package name}

\item{map}{remap model parameters from Stan to PKPDsim syntax, specified as a list. E.g. \code{map = list(V1 = V)}}

\item{drop}{vector of parameter names that we are not interested in or don't need to provide.}
}
\description{
Get prior for Torsten from a PKPDsim model package
}
