% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmc_posterior.R
\name{get_mcmc_posterior}
\alias{get_mcmc_posterior}
\title{Draw samples from the posterior given a Stan/Torsten model
and patient data}
\usage{
get_mcmc_posterior(
  mod,
  prior,
  regimen = NULL,
  covariates = NULL,
  data = NULL,
  nm_data = NULL,
  seed = 12345,
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{mod}{compiled Stan model}

\item{regimen}{drug regimen data created using \code{PKPDsim::new_regimen()}}

\item{covariates}{patient covariate data created using \code{PKPDsim::covariates()}}

\item{data}{data.frame with observed data.}

\item{nm_data}{NONMEM-style dataset. If \code{nm_data} supplied, then \code{regimen}, \code{covariates}, and \code{data} are not required.}

\item{seed}{seed for sampling}

\item{output_dir}{}

\item{...}{arguments passed to \code{mod$sample()}

Common arguments to cmdstanr include:
\itemize{
\item chains
\item parallel_chains
\item iter_warmup
\item iter_sampling
\item thim
\item adapt_delta
}}
}
\description{
Draw samples from the posterior given a Stan/Torsten model
and patient data
}
