% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Simple mode estimation function from density}
\usage{
get_mode(x, ...)
}
\arguments{
\item{x}{vector of continuous values for which to estimate mode}

\item{...}{function arguments passed on to \code{\link[stats:density]{stats::density()}}}
}
\description{
This function implements a simple estimation of model using the peak of the
univariate kernel density calculated by \code{\link[stats:density]{stats::density()}}. More advanced
methods are available in \code{modeest} package.
}
\keyword{internal}
