% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model_definitions.R
\name{get_n_cmt}
\alias{get_n_cmt}
\title{Get the number of compartments, either from analytical solver definition,
or from the dAdt definitions in the ODE block.}
\usage{
get_n_cmt(solver, ode = NULL)
}
\arguments{
\item{solver}{Torsten solver to use. Starts with \code{pmx_solve_}.}

\item{ode}{code for differential equation system, if not the analytic solver
is used.}
}
\description{
Get the number of compartments, either from analytical solver definition,
or from the dAdt definitions in the ODE block.
}
\keyword{internal}
