% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter_tables.R
\name{get_parameter_tables}
\alias{get_parameter_tables}
\title{Get parameter tables from posterior and prior}
\usage{
get_parameter_tables(post, params = NULL, long = FALSE)
}
\arguments{
\item{post}{posterior object from \code{get_mcmc_posterior()}}

\item{params}{vector of parameter names to extract from table with
posterior draws. If \code{NULL}, will take default parameters from \code{init} object.}
}
\description{
Get parameter tables from posterior and prior
}
