% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_quantile}
\alias{get_quantile}
\title{Get quantile for a given new value, given a vector of values.}
\usage{
get_quantile(x, y)
}
\arguments{
\item{x}{value to calculate quantile for}

\item{y}{vector of values, posterior draws in most cases}
}
\description{
Get quantile for a given new value, given a vector of values.
}
