% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Initialize a model
If model is not compiled yet, than compile it. Otherwise used the binary.}
\usage{
load_model(model, force = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{model}{model to be compiled, from the models available in the package in \code{inst/models}}

\item{force}{force recompile}

\item{verbose}{show output from Stan / cmdstanr? Defaults to \code{FALSE}.}

\item{...}{passed onto}
}
\description{
Initialize a model
If model is not compiled yet, than compile it. Otherwise used the binary.
}
