% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model_definitions.R
\name{parse_model_definition}
\alias{parse_model_definition}
\title{Parse model definitions into a list that can be used to generate a Stan
model}
\usage{
parse_model_definition(
  parameters,
  parameter_definitions,
  covariate_definitions,
  solver
)
}
\description{
Parse model definitions into a list that can be used to generate a Stan
model
}
