% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model_definitions.R
\name{parse_model_definitions}
\alias{parse_model_definitions}
\title{Parse model definitions into a list that can be used to generate a Stan
model}
\usage{
parse_model_definitions(
  parameters,
  parameter_definitions,
  covariate_definitions,
  solver,
  obs_types,
  n_theta,
  n_cmt,
  scale
)
}
\arguments{
\item{parameters}{list of parameter values, e.g. \code{list(CL = 5, V = 50)}.}

\item{parameter_definitions}{list of parameter definitions. See example
below.}

\item{covariate_definitions}{list of covariate definitions. See example
below.}

\item{solver}{Torsten solver to use. Starts with \code{pmx_solve_}.}

\item{obs_types}{vector of observed data types. If \code{NULL}, will use \code{"pk"}.}

\item{n_theta}{number of thetas. If analytic equation, will use number of
thetas for solver.}

\item{n_cmt}{number of compartments. If analytic equation, will use number
of compartments for solver.}

\item{scale}{scale definition. Can be either parameter expression or number.}
}
\description{
Parse model definitions into a list that can be used to generate a Stan
model
}
