% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKPDposterior.R
\docType{package}
\name{PKPDposterior-package}
\alias{PKPDposterior}
\alias{PKPDposterior-package}
\title{PKPDposterior: MCMC-Based MIPD}
\description{
MCMC-based inference based a population PK model and individual TDM data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightrx.github.io/PKPDposterior/}
}

}
\author{
\strong{Maintainer}: Ron Keizer \email{ron@insight-rx.com}

Authors:
\itemize{
  \item Jasmine Hughes \email{jasmine@insight-rx.com}
  \item Dominic Tong \email{dominic@insight-rx.com}
  \item Kara Woo \email{kara@insight-rx.com}
}

Other contributors:
\itemize{
  \item InsightRX [copyright holder, funder]
}

}
\keyword{internal}
