% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKPDsim_to_stan_data.R
\name{PKPDsim_to_stan_data}
\alias{PKPDsim_to_stan_data}
\title{Prepare data object for use in get_mcmc_posterior()}
\usage{
PKPDsim_to_stan_data(
  regimen,
  covariates,
  data,
  parameters,
  fixed = NULL,
  iiv,
  ruv,
  dose_cmt = 1,
  ltbs = FALSE,
  eiv_sd = NULL,
  verbose = FALSE
)
}
\arguments{
\item{regimen}{Regimen object (created by \code{\link[PKPDsim:new_regimen]{PKPDsim::new_regimen()}})}

\item{covariates}{List of covariate objects created by
\code{\link[PKPDsim:new_covariate]{PKPDsim::new_covariate()}}}

\item{data}{Data frame with columns \code{t}, \code{dv}, and \code{cmt}}

\item{parameters}{list of population parameters, e.g. \code{list(CL = 5, V = 50)}}

\item{fixed}{optional, vector of parameters to fix and not sample.}

\item{iiv}{list of inter-individual variability for parameters. Should have
exact same list elements as \code{parameters}, and magnitude supplied on SD scale.}

\item{ruv}{magnitude of residual unexplained variability (RUV). Should be a
list specifying proportional and/or additive error magnitude on standard
deviation scale, e.g. \code{list("prop" = 0.1, "add" = 1)}. If \code{ltbs} is TRUE,
should specify only an \code{add} part, which applies an additive error on the
log-scale (which then becomes an approximate proportional error).}

\item{dose_cmt}{Specify what dose compartment. Observation compartment in
dataset is irrelevant, handled in model.}

\item{ltbs}{use log-transform-both-sides approach for observations? Default
is \code{FALSE}.}

\item{eiv_sd}{vector of length 2 with sd for observation and
dose / other event types respectively (evid = 0, 1+).}

\item{verbose}{verbosity}
}
\description{
Prepare data object for use in get_mcmc_posterior()
}
\keyword{internal}
