% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_params}
\alias{plot_params}
\title{Plot posterior distributions of model parameters
If the model also supplies prior parameters, specified e.g. as \code{prior_CL}
for the \code{CL} parameter, it will plot them as well.}
\usage{
plot_params(post)
}
\arguments{
\item{post}{posterior object from \code{get_mcmc_posterior()}}
}
\value{
a ggplot2 object with posterior distributions (histograms) faceted
by parameter.
}
\description{
Plot posterior distributions of model parameters
If the model also supplies prior parameters, specified e.g. as \code{prior_CL}
for the \code{CL} parameter, it will plot them as well.
}
