% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percentile.R
\name{plot_percentile}
\alias{plot_percentile}
\title{Visual representation of percentile estimate}
\usage{
plot_percentile(x)
}
\arguments{
\item{x}{percentile of parameter or observation in posterior}
}
\description{
Adapted from PKPDmap::plot_eta()
}
