% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot predictions from posterior (or prior) simulations
Requires as input a summarized simulation data.frame, e.g. obtained using
\code{sim_from_draws(post, summarize=TRUE)}}
\usage{
plot_predictions(sim_data, obs = NULL)
}
\arguments{
\item{sim_data}{data.frame with simulated predictions}

\item{obs}{observed data. A data.frame with columns \code{t} and \code{dv}, e.g.
\code{tdm_data <- data.frame(t = c(2.5, 11.5),  dv = c(40, 14))}}
}
\description{
Plot predictions from posterior (or prior) simulations
Requires as input a summarized simulation data.frame, e.g. obtained using
\code{sim_from_draws(post, summarize=TRUE)}
}
