% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data}
\usage{
prepare_data(regimen, covariates, tdm_data)
}
\arguments{
\item{regimen}{Regimen object (created by \code{\link[PKPDsim:new_regimen]{PKPDsim::new_regimen()}})}

\item{covariates}{List of covariate objects created by \code{\link[PKPDsim:new_covariate]{PKPDsim::new_covariate()}}}

\item{tdm_data}{Data frame with columns t, dv, and cmt}
}
\value{
Named list suitable for passing on to Torsten
}
\description{
Prepare data
}
\examples{
regimen <- PKPDsim::new_regimen(interval = 12)
covariates <- list(
  WT = PKPDsim::new_covariate(value = 150, unit = "kg"),
  CRCL = PKPDsim::new_covariate(value = 6.5, unit = "l/hr")
)
tdm_data <- data.frame(t = c(1, 2), dv = c(900, 800), cmt = c(1, 1))
prepare_data(regimen, covariates, tdm_data)
}
