% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PKPDposterior.R
\name{print.PKPDposterior}
\alias{print.PKPDposterior}
\title{Print function for PKPDposterior object}
\usage{
\method{print}{PKPDposterior}(post)
}
\arguments{
\item{post}{posterior object (from get_mcmc_posterior())}
}
\description{
Print function for PKPDposterior object
}
