% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{remap}
\alias{remap}
\title{Remap named list or data.frame using a translation map}
\usage{
remap(x, map)
}
\arguments{
\item{x}{names list or data.frame}

\item{map}{map used to translate elements of x, e.g. \code{list("CLi" = "CL", "V" = "V1")}}
}
\description{
Remap named list or data.frame using a translation map
}
