% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKPDsim_to_stan_data.R
\name{tdm_to_nm}
\alias{tdm_to_nm}
\title{Convert TDM data to NONMEM format}
\usage{
tdm_to_nm(data)
}
\arguments{
\item{data}{Data frame with columns \code{t}, \code{dv}, and \code{cmt}}
}
\description{
Convert TDM data to NONMEM format
}
