% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stan_model.R
\name{write_stan_model}
\alias{write_stan_model}
\title{Write a Stan model code to file}
\usage{
write_stan_model(code, file = NULL)
}
\arguments{
\item{code}{Stan code (vector of character strings).}

\item{file}{filename to write to. If no name specified, will create temporary
file.}
}
\value{
filename of saved model
}
\description{
Write a Stan model code to file
}
