% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{covariates_to_nm}
\alias{covariates_to_nm}
\title{Convert list of covariate objects to NONMEM formatted data}
\usage{
covariates_to_nm(covariates)
}
\arguments{
\item{covariates}{List of covariate objects created by
\code{\link[PKPDsim:new_covariate]{PKPDsim::new_covariate()}}}
}
\description{
Convert list of covariate objects to NONMEM formatted data
}
