% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_map_estimates.R
\name{extract_map_estimates}
\alias{extract_map_estimates}
\title{Extra MAP estimates from the posterior}
\usage{
extract_map_estimates(post)
}
\arguments{
\item{post}{posterior estimate from \code{get_mcmc_posterior()}.}
}
\description{
Reports the mode for the posterior of each parameter
}
