% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_init.R
\name{get_init}
\alias{get_init}
\title{Get prior for Torsten from a PKPDsim model package}
\usage{
get_init(model)
}
\arguments{
\item{model}{package name}
}
\description{
Get prior for Torsten from a PKPDsim model package
}
