% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmc_posterior.R
\name{get_mcmc_posterior}
\alias{get_mcmc_posterior}
\title{Draw samples from the posterior given a Stan/Torsten model
and patient data}
\usage{
get_mcmc_posterior(mod, prior, data, seed = 12345, output_dir = getwd(), ...)
}
\arguments{
\item{mod}{compiled Stan model}

\item{data}{dataset (see \code{\link[=prepare_data]{prepare_data()}})}

\item{seed}{seed for sampling}

\item{output_dir}{}

\item{...}{arguments passed to \code{mod$sample()}

Common arguments to cmdstanr include:
\itemize{
\item chains
\item parallel_chains
\item iter_warmup
\item iter_sampling
\item thim
\item adapt_delta
}}
}
\description{
Draw samples from the posterior given a Stan/Torsten model
and patient data
}
\seealso{
\code{\link[=prepare_data]{prepare_data()}}
}
