% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmc_posterior.R
\name{get_mcmc_posterior}
\alias{get_mcmc_posterior}
\title{Draw samples from the posterior given a Stan/Torsten model
and patient data}
\usage{
get_mcmc_posterior(
  mod,
  init,
  data,
  seed = 12345,
  output_dir = tempdir(),
  verbose = TRUE,
  refresh = 0,
  ...
)
}
\arguments{
\item{mod}{compiled Stan model}

\item{init}{initial parameter set for sampler}

\item{data}{dataset (see \code{\link[=prepare_data]{prepare_data()}})}

\item{seed}{seed for sampling}

\item{output_dir}{output directory}

\item{verbose}{verbosity}

\item{refresh}{show output from sampler. Default is 0, meaning no output
from sampler is shown.}

\item{...}{arguments passed to \code{mod$sample()}

Common arguments to cmdstanr include:
\itemize{
\item chains
\item parallel_chains
\item iter_warmup
\item iter_sampling
\item thim
\item adapt_delta
}}
}
\description{
Draw samples from the posterior given a Stan/Torsten model
and patient data
}
\seealso{
\code{\link[=prepare_data]{prepare_data()}}
}
