% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmc_posterior.R
\name{get_mcmc_posterior}
\alias{get_mcmc_posterior}
\title{Draw samples from the posterior given a Stan/Torsten model
and patient data}
\usage{
get_mcmc_posterior(
  mod,
  data,
  seed = 12345,
  chains = 1,
  output_dir = tempdir(),
  verbose = TRUE,
  method = c("hmc", "vi"),
  skip_processing = FALSE,
  ...
)
}
\arguments{
\item{mod}{compiled Stan model}

\item{data}{dataset (see \code{\link[=prepare_data]{prepare_data()}})}

\item{seed}{seed for sampling}

\item{chains}{number of MCMC chains to simulate, passed on to Stan model}

\item{output_dir}{output directory}

\item{verbose}{verbosity}

\item{method}{either \code{hmc} (Hamilton Monte Carlo, using NUTS by default) or
\code{vi} (variational inference).}

\item{skip_processing}{will return the raw output from the sampler, without
further processing. This can sometimes be useful when errors occur in the
processing stage.}

\item{...}{arguments passed to \code{mod$sample()}

Common arguments to cmdstanr include:
\itemize{
\item chains
\item parallel_chains
\item iter_warmup
\item iter_sampling
\item thim
\item adapt_delta
}}
}
\description{
Draw samples from the posterior given a Stan/Torsten model
and patient data
}
\seealso{
\code{\link[=prepare_data]{prepare_data()}}
}
