% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Simple mode estimation function from density}
\usage{
get_mode(x)
}
\arguments{
\item{x}{vector of continuous values to estimate mode from}
}
\description{
More advanced methods are available in \code{modeest} package,
but likely not necessary for our purpose.
}
