% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter_tables.R
\name{get_parameter_tables}
\alias{get_parameter_tables}
\title{Get parameter tables from posterior and prior}
\usage{
get_parameter_tables(post, long = FALSE)
}
\arguments{
\item{post}{posterior object from \code{get_mcmc_posterior}}

\item{long}{return tables in long format? Default is \code{FALSE}}
}
\description{
Get parameter tables from posterior and prior
}
