% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Initialize a model
If model is not compiled yet, then compile it. Otherwise, use the binary.}
\usage{
load_model(model_file, force = FALSE, verbose = FALSE)
}
\arguments{
\item{model_file}{model file to be compiled, either full path to a model,
or else will look in \code{models} folder inside the installed package folder.}

\item{force}{force recompile}

\item{verbose}{show output from Stan / cmdstanr? Defaults to \code{FALSE}.}
}
\description{
Initialize a model
If model is not compiled yet, then compile it. Otherwise, use the binary.
}
