% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stan_model.R
\name{new_stan_model}
\alias{new_stan_model}
\title{Create a new stan model using a template}
\usage{
new_stan_model(
  parameters,
  parameter_definitions,
  covariate_definitions = NULL,
  solver = c("pmx_solve_twocpt", "pmx_solve_rk45"),
  template = "template.stan",
  obs_types = NULL,
  scale,
  n_cmt = 5,
  n_theta = NULL,
  file = NULL,
  verbose = FALSE
)
}
\description{
Create a new stan model using a template
}
