% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model_definitions.R
\name{parse_model_definitions}
\alias{parse_model_definitions}
\title{Parse model definitions into a list that can be used to generate a Stan
model}
\usage{
parse_model_definitions(
  parameters,
  parameter_definitions,
  ode,
  covariate_definitions,
  solver,
  obs_types,
  obs_cmt,
  scale,
  custom_ipred
)
}
\arguments{
\item{parameters}{list of parameter values, e.g. \code{list(CL = 5, V = 50)}.}

\item{parameter_definitions}{list of parameter definitions. See example
below.}

\item{ode}{code for differential equation system, if not the analytic solver
is used.}

\item{covariate_definitions}{list of covariate definitions. See example
below.}

\item{solver}{Torsten solver to use. Starts with \code{pmx_solve_}.}

\item{obs_types}{vector of observed data types. If \code{NULL}, will use \code{"pk"}.}

\item{obs_cmt}{observation compartment, or list of observation compartments
for each observation type when multiple observation types. E.g. \code{2} or
\code{list(pk = 2, pd = 8)}.}

\item{scale}{scale definition. Can be either parameter expression or number.}

\item{custom_ipred}{list of custom definitions for individual predictions for
each observation type. Sometimes this is needed when more flexibility is
needed than the \code{scale} parameter provides.}
}
\description{
Parse model definitions into a list that can be used to generate a Stan
model
}
