% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_from_draws.R
\name{sim_from_draws}
\alias{sim_from_draws}
\title{Simulate from posterior or prior draws using PKPDsim::sim()}
\usage{
sim_from_draws(post, model, prior = FALSE, n = NULL, ...)
}
\arguments{
\item{post}{posterior object from \code{get_mcmc_posterior()}}

\item{model}{PKPDsim model function}

\item{prior}{simulate from prior? Default \code{FALSE}, i.e. simulate from
posterior}

\item{n}{number of parameter draws / simulations to do. Defaults to \code{NULL},
i.e. use all draws.}
}
\description{
Simulate from posterior or prior draws using PKPDsim::sim()
}
