% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_stan_model.R
\name{validate_stan_model}
\alias{validate_stan_model}
\title{Validate the numerical simulation properties of a Stan model vs the model
written in PKPDsim.}
\usage{
validate_stan_model(
  stan_model,
  pkpdsim_model,
  data,
  parameters = NULL,
  mapping = NULL,
  n = 50,
  obs_types = c("pk", "pd"),
  max_abs_delta = 0.001,
  max_rel_delta = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{stan_model}{Stan/Torsten model}

\item{pkpdsim_model}{PKPDsim implementation}
}
\description{
Validate the numerical simulation properties of a Stan model vs the model
written in PKPDsim.
}
