% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_stan_model.R
\name{validate_stan_model}
\alias{validate_stan_model}
\title{Validate the numerical simulation properties of a Stan model vs the model
written in PKPDsim.}
\usage{
validate_stan_model(
  stan_model,
  pkpdsim_model,
  data,
  mapping = NULL,
  n = 50,
  max_abs_delta = 0.001,
  max_rel_delta = 1e-05,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{stan_model}{Stan/Torsten model}

\item{pkpdsim_model}{PKPDsim implementation}

\item{data}{dataset (see \code{\link[=new_stan_data]{new_stan_data()}})}

\item{mapping}{remap model parameters from Stan to PKPDsim syntax, specified
as a list. E.g. \code{map = list(V1 = V)}.}

\item{n}{number of patient datasets to use in validation. Default is 50.}

\item{max_abs_delta}{maximum allowed absolute delta between Stan and PKPDsim}

\item{max_rel_delta}{maximum allowed relative delta between Stan and PKPDsim,
in which PKPDsim is treated as reference.}

\item{verbose}{verbose output?}
}
\description{
Validate the numerical simulation properties of a Stan model vs the model
written in PKPDsim.
}
